♻️ refactor: 홈/인재 검색 인재 카드 통합 #413
- 로직과 ui 로직 상 중복되는 부분이 많은 홈/인재 검색에 사용되는 HomeEmployerEmployeeCardList와 EmployerEmployeeCardList에 사용되는 코드를 EmployeeCard로 통합했습니다.
