+++ src/components/Common/EmployeeCard/index.ts
@@ -1,0 +1,3 @@
+export { default as EmployeeCard } from '@/components/Common/EmployeeCard/EmployeeCard';
+export { default as EmployeeCardBookmark } from '@/components/Common/EmployeeCard/EmployeeCardBookmark';
+export { default as EmployeeCardEmptyState } from '@/components/Common/EmployeeCard/EmployeeCardEmptyState';
